NET-85: refactor network utils
